feat(unit-test): add a test for recovery with broken pipe
